requirements.txt
fastapi
uvicorn
neo4j
pydantic
pandas
python-multipart
python-dotenv
opensearch-py==2.7.1